Rework the creation of the crowbar client in chef
The "crowbar" system user does not have permission to read /etc/chef/webui.pem
by default. So call create client as root. And "knife configure" as "crowbar".

Addtionally the "crowbar" user on SUSE is currently created with the shell set to
"/sbin/nologin". sudo -i does not work in such a case. Call it with out it.